Rework `bindings!` macro
It's impossible to properly express "convert first element into Binding"
with a `macro_rules`. But I realized that we can just move some of the
things into a type system. This way it properly works with trailing
commas and no longer requires wrapping single elements in braces when
mixed with tuples.