Rework `bindings!` and `actions!` macro (#152)
* Rework `bindings!` macro

It's impossible to properly express "convert first element into Binding"
with a `macro_rules`. But I realized that we can just move some of the
things into a type system. This way it properly works with trailing
commas and no longer requires wrapping single elements in braces when
mixed with tuples.

* Use `related!` inside relationship macro

`RelationshipTarget::spawn` can supports up to 12 entities. This was
solved in https://github.com/bevyengine/bevy/pull/18865 To utilize the
fix, we need to use `related!`.