Clarify `Binding::MouseWheel`
Users sometimes expect it to capture vertical scroll, but it actually
outputs a `Vec2` with vertical scroll on the Y-axis. Originally, we
considered splitting the binding in Discord but decided to keep it as
is, since that's how Winit and Bevy represent it. I added a warning in
the docs, included a doc example, and extended the `simple_fly_cam`
example to showcase it.